Refactor SharedPreferences extension functions
